+++ 1.20/Forge/src/main/java/fuzs/puzzleslib/impl/init/ForgeGameRulesFactory.java
@@ -1,16 +1,11 @@
 package fuzs.puzzleslib.impl.init;
 
 import fuzs.puzzleslib.api.init.v2.GameRulesFactory;
-import fuzs.puzzleslib.mixin.accessor.BooleanValueForgeAccessor;
-import fuzs.puzzleslib.mixin.accessor.IntegerValueForgeAccessor;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.level.GameRules;
 
 import java.util.function.BiConsumer;
 
-/**
- * implementation on Forge, we need to use custom accessors as Forge has nothing to help us unfortunately
- */
 public final class ForgeGameRulesFactory implements GameRulesFactory {
 
     @Override
@@ -19,22 +14,12 @@ public final class ForgeGameRulesFactory implements GameRulesFactory {
     }
 
     @Override
-    public GameRules.Type<GameRules.BooleanValue> createBooleanRule(boolean defaultValue) {
-        return BooleanValueForgeAccessor.puzzleslib$callCreate(defaultValue, (minecraftServer, booleanValue) -> {});
+    public GameRules.Type<GameRules.BooleanValue> createBooleanRule(boolean defaultValue, BiConsumer<MinecraftServer, GameRules.BooleanValue> callback) {
+        return GameRules.BooleanValue.create(defaultValue, callback);
     }
 
     @Override
-    public GameRules.Type<GameRules.BooleanValue> createBooleanRule(boolean defaultValue, BiConsumer<MinecraftServer, GameRules.BooleanValue> changedCallback) {
-        return BooleanValueForgeAccessor.puzzleslib$callCreate(defaultValue, changedCallback);
-    }
-
-    @Override
-    public GameRules.Type<GameRules.IntegerValue> createIntRule(int defaultValue) {
-        return IntegerValueForgeAccessor.puzzleslib$callCreate(defaultValue, (minecraftServer, integerValue) -> {});
-    }
-
-    @Override
-    public GameRules.Type<GameRules.IntegerValue> createIntRule(int defaultValue, BiConsumer<MinecraftServer, GameRules.IntegerValue> changedCallback) {
-        return IntegerValueForgeAccessor.puzzleslib$callCreate(defaultValue, changedCallback);
+    public GameRules.Type<GameRules.IntegerValue> createIntRule(int defaultValue, int minimumValue, int maximumValue, BiConsumer<MinecraftServer, GameRules.IntegerValue> callback) {
+        return GameRules.IntegerValue.create(defaultValue, callback);
     }
 }